requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
dateparser==1.2.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
numpy==2.1.2
pandas==2.2.3
propcache==0.2.0
psycopg2==2.9.10
pycryptodome==3.21.0
python-binance==1.0.22
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
six==1.16.0
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
websockets==13.1
yarl==1.17.1
